+++ verifier/src/main/resources/com/aws/jverify/messages.properties
@@ -1,7 +1,6 @@
 compiler.err.pureMethodMultipleStatements=pure method should have only one statement
 compiler.err.pureMethodsNeedsReturnType=pure method should have a return type
 compiler.err.pureMethodNeedsReturnStatement=pure method statement should be a return
-compiler.err.multipleReturnNames=all postcondition lambda parameters must use the same name. Got {1} instead of {0}
 compiler.err.argumentMustBeLambda=the argument to a {0} call must be a lambda
 compiler.err.notSupported={0} is not supported
 compiler.err.contractAfterBody=call to JVerify header method {0} is not allowed after non-header statement
