+++ verifier/src/test/java/com/aws/jverify/verifier/tests/InterfacesVerification.java
@@ -6,7 +6,7 @@ import com.aws.jverify.testengine.JVerifyTest;
 import static com.aws.jverify.JVerify.*;
 
 @JVerifyTest(exitCode = 4, dafnyVerified = 10, dafnyErrors = 5)
-class Interfaces {
+class InterfacesVerification {
     public void root(I i) {
         var a = i.f(1);
 //              ^^^^^^ Error: function precondition could not be proved
@@ -116,9 +116,4 @@ class CInvalidImplementation implements I {
 //      ^^^ Error: assignment might update an object not in the enclosing context's modifies clause
         return 3;
     }
-}
-
-interface HasNoContractMethod {
-    @Pure
-    int hasNoContract();
 }
